feat: do the same for DateRangePicker
