+++ frontend/src/components/Calendar/CalendarBase/types.ts
@@ -4,3 +4,5 @@ export type CalendarBaseProps = Pick<
   UseProvideCalendarProps,
   'colorScheme' | 'isDateUnavailable' | 'monthsToDisplay'
 >
+
+export type DateRangeValue = [null, null] | [Date, null] | [Date, Date]
