+++ frontend/src/components/Calendar/index.ts
@@ -1,2 +1,3 @@
 export * from './Calendar'
+export type { DateRangeValue } from './CalendarBase/types'
 export * from './RangeCalendar'
